Add #9 React useState Hook with TypeScript demonstration
